Extract editor command transaction helpers
